requirements.txt
plaintext
#aenum==3.1.15
#aiohttp==3.9.3
#aiosignal==1.3.1
#astroid==3.0.3
#attrs==23.2.0
#bcrypt==4.1.2
#black==24.1.1
#boto3==1.34.39
#botocore==1.34.39
#certifi==2024.2.2
#cffi==1.16.0
#charset-normalizer==3.3.2
#click==8.1.7
#cryptography==42.0.2
#dill==0.3.8
#frozenlist==1.4.1
#greenlet==3.0.3
#idna==3.6
#iniconfig==2.0.0
#isort==5.13.2
#jmespath==1.0.1
#Mako==1.3.2
#MarkupSafe==2.1.5
#mccabe==0.7.0
#multidict==6.0.5
#mypy-extensions==1.0.0
#mysql-connector-python==8.3.0
#numpy==2.1.0
#packaging==23.2
#pandas==2.2.2
#paramiko==3.4.0
#parse==1.20.1
#parse-type==0.6.2
#parsedatetime==2.6
#pathspec==0.12.1
#platformdirs==4.2.0
#playwright==1.46.0
#pluggy==1.5.0
#pycparser==2.21
#pyee==11.1.0
#pylint==3.0.3
#PyNaCl==1.5.0
#pytest>=7.0.0,<8.0.0
#pytest-base-url==2.1.0
#pytest-bdd~=4.0.2
#pytest-lazy-fixture==0.6.3
#pytest-ordering==0.6
#pytest-playwright==0.4.4
#pytest-reportportal==5.3.2
#python-dateutil==2.8.2
#python-decouple==3.8
#python-slugify==8.0.4
#pytz==2024.1
#PyYAML==6.0.1
#reportportal-client==5.5.4
#requests==2.31.0
#ruamel.yaml==0.18.6
#ruamel.yaml.clib==0.2.8
#s3transfer==0.10.0
#six==1.16.0
#text-unidecode==1.3
#tomlkit==0.12.3
#typing_extensions==4.9.0
#tzdata==2024.1
#urllib3==2.0.7
#utils==1.0.2
#wait-for==1.2.0
#yarl==1.9.4


black~=24.3.0
pylint~=3.0.3
boto3~=1.34.39
botocore~=1.34.39
mysql-connector-python==9.2.0
paramiko==3.4.0
playwright~=1.46.0
pytest==7.4.4
pytest-bdd~=4.0.2
pytest-lazy-fixture==0.6.3
pytest-ordering==0.6
pytest-reportportal==5.4.1
python-decouple==3.8
pytest-playwright==0.4.4
PyYAML==6.0.1
ruamel.yaml==0.18.6
wait-for==1.2.0
setuptools~=69.1.0
jmespath~=1.0.1
pip~=24.2
attrs~=23.2.0
wheel~=0.41.2
cryptography~=44.0.1
utils~=1.0.2
typing_extensions~=4.9.0
dill~=0.3.8
numpy~=2.1.0
Mako~=1.3.2
MarkupSafe~=2.1.5
pyee~=11.1.0
idna~=3.7
multidict~=6.0.5
aenum~=3.1.15
platformdirs~=4.2.0
tomli~=2.0.1
packaging~=23.2
pathspec~=0.12.1
click~=8.1.7
s3transfer~=0.10.0
isort~=5.13.2
aiohttp~=3.10.11
astroid~=3.0.3
tomlkit~=0.12.3
mccabe~=0.7.0
pluggy~=1.5.0
iniconfig~=2.0.0
py~=1.11.0
yarl~=1.9.4
aiosignal~=1.3.1
frozenlist~=1.4.1
certifi~=2024.7.4
python-slugify~=8.0.4
python-dateutil~=2.8.2
urllib3~=2.2.2
six~=1.16.0
greenlet~=3.0.3
bcrypt~=4.1.2
parsedatetime~=2.6
pycparser~=2.21
parse~=1.20.1
requests~=2.32.0
env~=0.1.0
pandas~=2.2.2
python-decouple>=3.5